requirements.txt
fastapi
sqlmodel
uvicorn
python-dotenv
pydantic-settings
uuid
psycopg2-binary
python-jose[cryptography]
passlib[bcrypt]
python-multipart
bcrypt==4.0.1